Add files via upload
I have constructed a predictive model using the Titanic dataset to determine the likelihood of a passenger surviving the Titanic disaster based on three key features: gender, age, and class. The model is built upon logistic regression and has been trained with the data available. By providing input values for gender (coded as Male: 1, Female: 0), age, and class (1, 2, or 3), the model outputs a prediction indicating whether the passenger is likely to have survived or not.